Subscript collapsing (#25)
This PR implements collapsing subscripts in presentation MathML
documents. To get this to work, a number of other refactoring steps were
needed, the most significant of which is to convert the arguments of the
enum variants in `MathExpression` from `&'a str` to `String` in order to
have them own the data themselves rather than just pointers to it. This
makes traversal and in-place modification much easier, at the cost of
introducing `.to_string()` method calls in the test functions.

Unnormalized presentation MathML expression:


![original_sir](https://user-images.githubusercontent.com/653549/200692198-1a691a04-2ca2-4922-95ef-f9ff8700a920.png)


Normalized expression:

![sir_collapsed_subscripts](https://user-images.githubusercontent.com/653549/200692265-d2ed6092-bda2-4f19-a2ff-5a0e6809c3cc.png)

The implementation is a minimal, proof of concept one, so it may not be
able to handle more complicated expressions, but at least it works with
the CHIME SIR update equation.

Co-authored-by: Adarsh Pyarelal <adarsh@arizona.edu>