Take another pass (or ten) over mention linking (#55)
Here's a draft showing how I understand the mention linking code. AFAICT
it produces the same output file as the previous version does. Further
changes would probably start to deviate from the original, so this is a
good time to take stock. This version is using a debug time stamper and
uid stamper so that the output stays consistent. I recall the hash
function for the uid strangely producing different output on the same
input. There are several TODOS:

* escape the names
* add line numbers to comments that don't have them
* adjust dynamically for programming language
* actually deserialize all the json
* update the Scala version :-)
* improve the alignment