Equation reading updates (#75)
This PR implements a number of updates to the equation reading component
of SKEMA.

- The old LSTM-based model has been updated to a new transformer-based
model (cnn_xfmer_OMML-90K_best_model_RPimage.pt) with a much higher BLEU
score of 95.6.
- The `im2mml` code has been refactored to be part of the top-level
`skema` package.
- The contents of `generate_mathml` have been brought up one level.
- There is a new module, `skema/im2mml/api.py` that contains
public-facing functions that are also used by the Im2MML webservice.
- A unit test for the equation reading pipeline has been added. You can
run it by invoking `pytest` in the `skema/im2mml` directory. Currently,
this test runs fine locally, but for some reason is getting stuck when
running on the CI server. Will have to look into that later.

Co-authored-by: gauravsh0812 <gauravs@arizona.edu>
Co-authored-by: Adarsh Pyarelal <adarsh@arizona.edu>