Show how to use Pdf2Txt (#56)
Output is

```
   rawText: fi gures
cookedText: figures .

   rawText: o ffi cials
cookedText: officials .

   rawText: di ffi cult
cookedText: difficult.

   rawText: sta ffi ng
cookedText: staffing .

   rawText: shif ts
cookedText: shifts .

   rawText: speci fi cally
cookedText: specifically .

   rawText: ca not
cookedText: canot .

   rawText: we ' ve
cookedText: we 've .

   rawText: tra ffi c
cookedText: traffic .

   rawText: fi rst
cookedText: first .

   rawText: Here 's
cookedText: Here 's.

   rawText: Let 's
cookedText: Let 's.

   rawText: hospital s
cookedText: hospitals .
```

The apostrophes are not corrected and they might need a special case.
hospitals and shifts are corrected by the WordBreakBySpacePreprocessor,
which is not normally used. It often overcorrects, as it does here for
canot. I'm not sure that `shif ts` is a normal ligature problem. Usually
the f and t would be together, but it could be added to the list. It
would probably be better to have the mistakes accompanied by their
contexts.

Co-authored-by: Enrique Noriega <enrique.noriega@gmail.com>