Assemble for text reading (#106)
This addition *should* make it possible to produce a fat jar file. We
need to fill in the `mainClass` at the very least. It should be tested,
even if unit tests are passing. Sometimes people have tried to access
files that are no longer available outside the development environment
and it only shows up now. Plus I might not have merged some of the files
correctly. Also, it might be good not to compress the large resource
files as they are added to the jar. It takes a long time and then they
need to be repeatedly decompressed each time the program runs, slowing
startup time. I can add that later after everything is known to work.