Scenario Context (#117)
This is the first implementation of Scenario Context.

The two main abstractions are: `ContextEngine`, which looks around a
window of sentences in order to attach location and time scenario
context to relation and event mentions; and `MentionOrderer` which is
used to abstract the notion of a linear partial ordering of appearance
for the mentions, regardless of the source. This way, we can sort the
mentions that come from a Cosmos document, which has pages and blocks
within pages, and look around efficiently at the time of context
resolution.