Addresses Built-in Map Generation (#173)
This PR addresses an issue with how the Python built-in map is created.

Before, the built-in map stored in
`skema/program_analysis/PyAST2CAST/python_builtins.yaml` couldn't be
loaded on the fly. This was due to an incorrect way of trying to load
the yaml file.
This has been fixed now, and the yaml file can be loaded appropriately
using the "__file__" global variable.

In addition, a simple test case was created to test the functionality.