Implement cosmosJsonToMentions for webapp (#214)
It seems to work through the browser and there is also a unit test that
indicates as much, but I don't know the details of the file format to
trust it for sure. In the OpenAPI file I've left TODOs for all the
descriptions and only used placeholders for the input and output
formats. I'd appreciate pointers to any specification available. These
look like very large structures.

This is for #210.

---------

Co-authored-by: Enrique Noriega <enoriega@arizona.edu>