Morae demo 6/28 (#269)
This PR adds an additional put endpoint to the skema-rs morae api that
supports a transactional workflow of the FN -> AMR pipeline, with
support for SIR/PN dynamics. The model is converted and pushed to the
DB, pulled from DB and converted to an AMR, and then deleted from DB.

It also includes the demo notebook for the 6/28 open house and
corresponding data for it as well.