Unary Operator Wire Fix PR (#314)
This PR addresses an issue with wiring when Unary operations (i.e.
UnarySub/Minus and the rarer UnaryAdd). The wires were not being
generated correctly and were returning negative indexing. This has now
been fixed with a rewrite of handling the Unary Operations.

This PR also introduces some initial tests for the expression files that
check properties and correctness of the Gromets.

Resolves #200 
Resolves #268