Support for parsing MathML with light interpretation (#326)
This PR implements support for parsing presentation MathML in general
with a small amount of preliminary interpretation (e.g., Globbing `S(t)`
to produce the identifier `S`. Previously, this functionality was
restricted to first order ODE parsing, but has been extended to parsing
arbitrary expressions as well. This is the default behavior when
`MathExpressionTree` structs are constructed from strings. The
interpreting parsers are in `interpreted_mathml.rs`

The previous 'non-interpreting' parsers are still available in
`generic_mathml.rs`, and are kept for compatibility with ISA code. Some
of them are also reused in `interpreted_mathml.rs`

---------

Co-authored-by: Adarsh Pyarelal <adarsh@arizona.edu>