Introduces changes.md to document semver tags (#425)
## Summary of changes

### `changes.md`
This PR introduces a new page in our project site that documents changes
associated with each semver tag we create. At the end of each sprint,
we'll create a new entry in our log of changes that documents a new
semver tag incrementing the last patch version. Pushing this tag will
trigger a series of docker images to be published with the new semver
tag.

### PR auto-labeling workflow
As defined in `.github/labeler.yml` and
`.github/workflows/add-labels-to-pr.yml`, PRs are now auto-tagged to
highlight the modified system component based on what files were changed
in the PR.


Resolves #422