[Code2FN] GroMEt Function Call Handler Rewrite PR (#483)
### Refactoring of GroMEt Function Call Handler
- The handler that creates the GroMEt elements related to function calls
needed a rewrite due to it becoming very difficult to maintain and
change when bugs were discovered.
- The main goal was to make the code significantly smaller, and to
remove redundancy and streamline the process as much as possible.
- The biggest redundancy changes was with the argument handling, as it
was being done an unnecessary amount of times.
- There was also inconsistency with pof port generation and then it was
needed and not needed. This caused inconsistent wiring generation.
- A test file was added to support checking Gromets of small example
programs that exercise the generation of function calls for both normal
and primitive function calls.

### Removing Docstrings from Python AST to CAST
- Python stores multi-line comments as docstrings. These don't get
eliminated from the AST, as they get stored as normal strings.
- This causes some issues when creating the CAST, because they are left
in there as floating strings that don't contribute much to the
extraction from this perspective. They're more relevant when they're
extracted as comments.
- I've gone ahead and added code that removes all instances of
docstrings that appear at the function definition level, global level,
and class definition level. Now there'll no longer be any floating
docstrings in the CAST.
- The only strings that remain in the CAST are ones that are used as an
actual value in a computation (assignment, function call, etc..)

### Additional Changes
- Removing commented out lines of code that aren't needed.
- Minor formatting fixes done with Black autoformatter.

Resolves #406 
Resolves #380 
Resolves #197 
Resolves #434