MATLAB Tree-sitter regression testing with GitHub CI (#504)
## MATLAB Tree-sitter regression testing with GitHub CI

This PR demonstrates key elements of an agile test plan supporting the
buildout of the MATLAB grammar within Tree-sitter. The intent is to
create a regression testing system that is both easy to modify and easy
to use. Tests are performed autonomously by GitHub on main branch
commit, and can be run locally.

## Relevant features:
- Use of pytest within GitHub Workflows.
- Use of CAST elements within pytest.
- Data files in different directories than the tests.
- Use of n data file tests within one test file.
- Software design in the same style as existing SKEMA CI tests.

## Related issues:

- Completes issue #501
- Partially completes issue #420

---------

Co-authored-by: Joseph Astier <jastier@arizona.edu>
Co-authored-by: Gus Hahn-Powell <ghp@lum.ai>