[Code2FN] Gromet FN Missing Function Types Fix (#575)
This PR contains a couple of changes to the Gromet generation to fix
certain Gromet FN call boxes not containing what kind of function type
they were.

### Missing function_type Fix
This issue manifested when the function call was type 'Attribute' for a
variable that contains an object which we don't have any real
information about. In this case, the Function Type should be
"IMPORTED_METHOD" with an import type of "OTHER". Some additional code
was added to handle this case appropriately.

### Missing import_method Fix
This issue manifested when a function call with type 'Attribute' for a
Python library was done. As an example: calls from the numpy library
were missing the function_type and the import_type. Some additional
handling was added to support this.


Resolves #537 
Resolves #581