[execution_engine] Endpoint for parameter extraction and amr enrichment (#608)
## AMR enrichment endpoint
Creates an endpoint `amr-enrich` which accepts an AMR Dict, a source
file string, and a filename. The endpoint uses the execution engine to
extract values for the parameters listed in the AMR and returns an
enriched version with the `value` field filled in.

### Testing
Creates test_execution_server.py for unit tests relating to
execution_engine endpoints

### Other changes
- Improved error handling for execution engine allows
- Fixes bug where path to Gromet not being properly resolved
- Adds additional Gromet preprocessing step before Memgraph ingestion
- Fixes floating point precision bug
-  Adds :Opi as possible node type for assignment right hand side

---------

Co-authored-by: Gus Hahn-Powell <ghp@lum.ai>