Model Coverage Regression Testing PR (#694)
This PR adds some modifications and a test case to help us determine
when we've had a regression in model coverage. This is part of the
effort to develop more tests to maintain stability within the Code2FN
pipeline.

## Summary of Changes
- Adds a test script (test_model_coverage.py) that checks the coverage
against all the current models we support, it also checks to make sure
that whenever we run the test pipeline that we generate reports for all
the models that we support.
- Modified model_coverage_report.py to output an additional JSON file
that contains the 'raw' model coverage information (i.e. the
supported/total lines) that is used by the test script to check for
regressions in model coverage.


Resolves #677

---------

Co-authored-by: Vincent Raymond <vincent@lum.ai>