Python tree-sitter to CAST porting: Loops (#745)
This PR introduces support for generating CAST for Loops (for/while)
using tree-sitter, as part of the ongoing effort to port over the Python
AST to CAST generation to using tree-sitter.

### Python Tree Sitter
- Added support for generating CAST Loop nodes using Python tree sitter.
- Specifically, we added support for Python's For and While loop CAST
generation.
- Added some support for tree-sitter "patterns: "list_pattern",
"tuple_pattern", "list_pattern". These are used primarily in the For
loop syntax as the item we're iterating over.
- Also added support for generating Iterator function calls, as used by
the Python For Loops.
- Updated the CAST to AGraph visualizer to better visualize tuples.

### Testing 

- Added some small unit tests for loops (for/while) to determine
consistency.
- Added a small unit test for detecting missing identifiers in the first
line of a Python program.

### Other Fixes
- Fixes an issue with missing identifier names when they appeared on the
first line of the Python program. This was done by adding some
additional handling that is specific to the first line of the program.

Resolves #498 
Resolves #740