MATLAB to GroMEt testing (#765)
## MATLAB to GroMEt testing
This PR adds GroMEt translation to all pytest cases in the Tree-Sitter
MATLAB translator. This was done as part of the fix for a bug that was
happening when generating Model Coverage Reports. The CAST to GroMEt
converter was reporting correctly that the "list" type is not supported,
switching to the corresponding enumeration fixed that issue.

## Relavant Features:
- Each pytest case now includes GroMEt translation.
- Got rid of some outdated static method typing calls.
- `LiteralValue `data types are now CAST enumerations instead of
hard-coded strings.
- Removed the handler for `Range `Tree-sitter syntax nodes, these nodes
are managed by iterator routines
- Turned off the pytest series related to loops temporarily

Related to issue #725 
Partially resolves issue #764

---------

Co-authored-by: Joseph Astier <jastier@arizona.edu>
Co-authored-by: Tito Ferra <tito@lum.ai>
Co-authored-by: titomeister <ferra@email.arizona.edu>