Python tree-sitter to CAST porting: Classes (#850)
This PR introduces support for generating CAST for Python Classes. In
CAST the translation for these is RecordDef. This is part of the ongoing
effort of porting over the Python to CAST generation to use tree-sitter
instead of the Python AST.

### Summary of Changes
- Adds support in python/ts2cast.py for generating RecordDef CAST nodes,
which represent Python classes.
- Adds test script test_record_cast.py for maintaining consistency with
basic examples.
- Adds a handler of CAST Name nodes for when they're being used in
Operators or returns (get_operand_node). This is to make sure the
correct CAST node is being used in these at all times.
- Very minor code cleanup.


Resolves #805

---------

Co-authored-by: Vincent Raymond <vincent@lum.ai>